Updated integrators to take a long paramter for the number of steps to run and return a long for the number of steps that were actually completed
git-svn-id: http://svn.code.sf.net/p/protomol/svn/trunk@2084 ce46b6ef-3445-0410-911d-c47028e6f9df

+++ protomol/src/protomol/ProtoMolApp.h
@@ -52,8 +52,8 @@ namespace ProtoMol {
     GenericTopology *topology;
 
     // Run
-    int currentStep;
-    int lastStep;
+    long currentStep;
+    long lastStep;
 
     ProtoMolApp() {}
     ProtoMolApp(ModuleManager *modManager);
